Adding commit status on export task to external file
